+++ MobileRTCSample/MobileRTCSample/MeetingSettingsViewController.h
@@ -8,6 +8,10 @@
 
 #import <UIKit/UIKit.h>
 
+#define Custom_Meeting_Cell_Tag 10000
+#define Raw_Data_Cell_Tag 10001
+#define Raw_Data_UI_Enable @"raw.data.ui.enable"
+
 @interface MeetingSettingsViewController : UITableViewController
 
 @end