+++ MobileRTCSample/MobileRTCSample/SDKPresenters/how_to_use_inmeeting_function/meeting_callback/SDKStartJoinMeetingPresenter+AudioServiceDelegate.m
@@ -29,4 +29,11 @@
     }
 }
 
+#if 0
+- (void)onSinkMeetingAudioRequestUnmuteByHost
+{
+    NSLog(@"the host require meeting attendants to enable microphone");
+}
+#endif
+
 @end