+++ MobileRTCSample/MobileRTCSample/SDKPresenters/other_features/schedule_meeting_for_login_user/SDKScheduleMeetingPresenter.m
@@ -10,7 +10,6 @@
 
 @implementation SDKScheduleMeetingPresenter
 
-#warning  TODO edit meeting
 
 - (BOOL)scheduleMeeting:(nonnull id<MobileRTCMeetingItem>)meetingItem WithScheduleFor:(nullable NSString *)userEmail
 {